+++ Presupuestos/Controllers/EmpleadosController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using Presupuestos.Models;
+using Presupuestos.Comun;
 
 namespace Presupuestos.Controllers
 {
@@ -17,6 +18,14 @@ namespace Presupuestos.Controllers
         // GET: /Empleados/
         public ActionResult Index()
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             var tblempleados = db.TblEmpleados.Include(t => t.TblPuesto);
             return View(tblempleados.ToList());
         }
@@ -24,6 +33,14 @@ namespace Presupuestos.Controllers
         // GET: /Empleados/Details/5
         public ActionResult Details(int? id)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -39,6 +56,14 @@ namespace Presupuestos.Controllers
         // GET: /Empleados/Create
         public ActionResult Create()
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             ViewBag.IDPuesto = new SelectList(db.TblPuestos, "IDPuesto", "Descripcion");
             return View();
         }
@@ -50,6 +75,14 @@ namespace Presupuestos.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include="IDEmpleado,Nombre,Paterno,Materno,IDPuesto")] TblEmpleado tblempleado)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             if (ModelState.IsValid)
             {
                 db.TblEmpleados.Add(tblempleado);
@@ -64,6 +97,14 @@ namespace Presupuestos.Controllers
         // GET: /Empleados/Edit/5
         public ActionResult Edit(int? id)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -84,6 +125,14 @@ namespace Presupuestos.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include="IDEmpleado,Nombre,Paterno,Materno,IDPuesto")] TblEmpleado tblempleado)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             if (ModelState.IsValid)
             {
                 db.Entry(tblempleado).State = EntityState.Modified;
@@ -97,6 +146,14 @@ namespace Presupuestos.Controllers
         // GET: /Empleados/Delete/5
         public ActionResult Delete(int? id)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -114,6 +171,14 @@ namespace Presupuestos.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             TblEmpleado tblempleado = db.TblEmpleados.Find(id);
             db.TblEmpleados.Remove(tblempleado);
             db.SaveChanges();
