+++ Presupuestos/Controllers/EmpresaController.cs
@@ -24,6 +24,9 @@ namespace Presupuestos.Controllers
             if (user == null)
                 return Redirect("~/Default");
 
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             return View(db.TblEmpresas.ToList());
         }
 
@@ -34,6 +37,9 @@ namespace Presupuestos.Controllers
 
             if (user == null)
                 return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
 
             if (id == null)
             {
@@ -55,6 +61,9 @@ namespace Presupuestos.Controllers
             if (user == null)
                 return Redirect("~/Default");
 
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             return View();
         }
 
@@ -69,6 +78,9 @@ namespace Presupuestos.Controllers
 
             if (user == null)
                 return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
 
             if (ModelState.IsValid)
             {
@@ -87,6 +99,9 @@ namespace Presupuestos.Controllers
 
             if (user == null)
                 return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
 
             if (id == null)
             {
@@ -112,6 +127,9 @@ namespace Presupuestos.Controllers
             if (user == null)
                 return Redirect("~/Default");
 
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             if (ModelState.IsValid)
             {
                 db.Entry(tblempresa).State = EntityState.Modified;
@@ -128,6 +146,9 @@ namespace Presupuestos.Controllers
 
             if (user == null)
                 return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
 
             if (id == null)
             {
@@ -150,6 +171,9 @@ namespace Presupuestos.Controllers
 
             if (user == null)
                 return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
 
             TblEmpresa tblempresa = db.TblEmpresas.Find(id);
             db.TblEmpresas.Remove(tblempresa);
