+++ Presupuestos/Controllers/ProyectosController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using Presupuestos.Models;
+using Presupuestos.Comun;
 
 namespace Presupuestos.Controllers
 {
@@ -17,6 +18,14 @@ namespace Presupuestos.Controllers
         // GET: /Proyectos/
         public ActionResult Index()
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             var tblproyectos = db.TblProyectos.Include(t => t.TblCliente).Include(t => t.TblEstatu);
             return View(tblproyectos.ToList());
         }
@@ -24,6 +33,14 @@ namespace Presupuestos.Controllers
         // GET: /Proyectos/Details/5
         public ActionResult Details(int? id)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -39,6 +56,14 @@ namespace Presupuestos.Controllers
         // GET: /Proyectos/Create
         public ActionResult Create()
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             ViewBag.IDCliente = new SelectList(db.TblClientes, "IDCliente", "NombreComercial");
             ViewBag.IDEstatus = new SelectList(db.TblEstatus, "IDEstatus", "Descripcion");
             return View();
@@ -51,6 +76,14 @@ namespace Presupuestos.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include="IDProyecto,IDCliente,Nombre,Decripcion,FechaInicio,FechaFin,Activo,IDEstatus")] TblProyecto tblproyecto)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             if (ModelState.IsValid)
             {
                 db.TblProyectos.Add(tblproyecto);
@@ -66,6 +99,14 @@ namespace Presupuestos.Controllers
         // GET: /Proyectos/Edit/5
         public ActionResult Edit(int? id)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -88,6 +129,14 @@ namespace Presupuestos.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include="IDProyecto,IDCliente,Nombre,Decripcion,FechaInicio,FechaFin,Activo,IDEstatus")] TblProyecto tblproyecto)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             if (ModelState.IsValid)
             {
                 db.Entry(tblproyecto).State = EntityState.Modified;
@@ -102,6 +151,14 @@ namespace Presupuestos.Controllers
         // GET: /Proyectos/Delete/5
         public ActionResult Delete(int? id)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -119,6 +176,14 @@ namespace Presupuestos.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
             TblProyecto tblproyecto = db.TblProyectos.Find(id);
             db.TblProyectos.Remove(tblproyecto);
             db.SaveChanges();
@@ -128,6 +193,14 @@ namespace Presupuestos.Controllers
 
         public ActionResult ActualizaDetalleProyecto(int id) 
         {
+            var user = Session["User"] as mUsuario;
+
+            if (user == null)
+                return Redirect("~/Default");
+
+            if (user.Tipo.Equals(2))
+                return Redirect("~/Presentacion/FrmHomeEmpleados.aspx"); 
+
              dbPresupuestosEntities db = new dbPresupuestosEntities();
 
              if (id == null)
